api/requirements.txt
fastapi 
uvicorn[standard]
SQLAlchemy
mariadb 
python-dotenv
pydantic[email]
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytest